+++ src/utilities.ts
@@ -2,8 +2,6 @@ import { djinnjsOutDir } from "./config";
 
 /**
  * Appends resources to the documents head if it hasn't already been loaded.
- * @param filenames - a filename `sting` or an array of `string` CSS filenames or a URL -- exclude the extension
- * @deprecated will be removed in v1.0.0 - use the `djinnjs/utilities` version
  */
 export function fetchCSS(filenames: string | Array<string>): Promise<{}> {
     return new Promise(resolve => {
@@ -47,4 +45,13 @@ export function fetchCSS(filenames: string | Array<string>): Promise<{}> {
             }
         }
     });
+}
+
+/**
+ * Safely register Web Components to Custom Element Registry.
+ */
+export function mount(tagName: string, className: CustomElementConstructor): void {
+    if (!customElements.get(tagName)) {
+        customElements.define(tagName, className);
+    }
 }